backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.127.1",
    "google-genai>=1.56.0",
    "modal>=1.3.0.post1",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlmodel>=0.0.30",
    "uvicorn>=0.40.0",
]
